# ecommerce/docs/swagger.yaml
definitions:
...
  controllers.productInput:
    properties:
      cost_price:
        type: integer
      image_url:
        type: string
      keterangan:
        type: string
      nama:
        type: string
      product_type_id:
        type: integer
      qty_available:
        type: integer
      qty_forecasted:
        type: integer
      uom_id:
        type: integer
    type: object
  controllers.productTypeInput:
    properties:
      type:
        type: string
    type: object
...
  models.Product:
    properties:
      cost_price:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      keterangan:
        type: string
      nama:
        type: string
      product_type_id:
        type: integer
      qty_available:
        type: integer
      qty_forecasted:
        type: integer
      uom_id:
        description: foreign key
        type: integer
      updated_at:
        type: string
    type: object
  models.ProductType:
    properties:
      created_at:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /product:
    get:
      description: Get a list of Product.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get all Product.
      tags:
      - Product
    post:
      description: Creating a new Products. (Admin Only)
      parameters:
      - description: the body to create a new Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.productInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      security:
      - BearerToken: []
      summary: Create New Products.
      tags:
      - Product
  /product-type:
    get:
      description: Get a list of Product Type.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProductType'
            type: array
      summary: Get all Product Types.
      tags:
      - Product Type
    post:
      description: Creating a new Product Type.
      parameters:
      - description: the body to create a new ProductType
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.productTypeInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductType'
      security:
      - BearerToken: []
      summary: Create New Product Types.
      tags:
      - Product Type
  /product-type/{id}:
    delete:
      description: Delete a Product Type by id.
      parameters:
      - description: product type id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Product Types.
      tags:
      - Product Type
    put:
      description: Update Product Type by id.
      parameters:
      - description: ProductType id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update ProductType
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.productTypeInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductType'
      security:
      - BearerToken: []
      summary: Update Product Types.
      tags:
      - Product Type
  /product/{id}:
    delete:
      description: Delete a Product by id. (Admin Only)
      parameters:
      - description: so id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Product.
      tags:
      - Product
    get:
      description: Get a Product by id.
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Get a Product.
      tags:
      - Product
    put:
      description: Update Product by id. (Admin Only)
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.productInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      security:
      - BearerToken: []
      summary: Update Products.
      tags:
      - Product
...
swagger: "2.0"